+++ src/register/register-config.ts
@@ -1,16 +1,12 @@
 import * as vscode from 'vscode';
 import * as fs from 'fs';
 import * as path from 'path';
-import { execSync } from 'child_process';
-import beforePluginInit from "../module/mixin-beforePluginInit";
+import beforePluginInit from '../module/mixin-beforePluginInit';
 import NotificationService from '../utils/notificationService';
-import { resolveResult } from '../utils/promiseResolve';
-import { ignoreArray, generateKeywords, overwriteIgnoreFilesLocally, isGitTracked } from '../utils/index';
+import onPluginInit from "../module/mixin-onPluginInit";
+import { ignoreArray, generateKeywords } from '../utils/index';
 import { MergeProperties, properties, computeGitChanges } from '../global-object/properties';
-import EventBus from '../utils/emitter';
-
-const CONFIG_FILES = properties.configFileSchema;
-type ConfigFile = (typeof CONFIG_FILES)[number];
+import { CONFIG_FILES, type ConfigFile } from '../types/Properties';
 
 // 通用的配置读取
 async function readConfigFile(uri: vscode.Uri): Promise<any | null> {
@@ -91,7 +87,7 @@ async function handleConfig(uri: vscode.Uri, context?: vscode.ExtensionContext) 
     // 否则为 true
     vscode.commands.executeCommand('setContext', 'Extension.logrcNotFound', !logrcFound);
     if (config) {
-      initPlugins(config);
+      onPluginInit(config);
     }
   } catch (err) {
     console.log('err', err);
@@ -155,52 +151,10 @@ export async function registerConfig(context: vscode.ExtensionContext) {
   if (!fs.existsSync(configPath)) {
     MergeProperties({ configResult: false });
     vscode.commands.executeCommand('setContext', 'Extension.logrcNotFound', true);
-    initPlugins('.logrc');
+    onPluginInit('.logrc');
   }
   // 读取每个工作区的配置文件
   registerConfigWatchers(context);
   // 分开调用，逻辑更清晰
   registerSaveWatcher(context);
-}
-
-function initPlugins(config: ConfigFile) {
-  switch (config) {
-    case '.gitignore':
-      return setIgnoredFiles();
-    case '.logrc':
-      return setLogrc();
-    default:
-      return;
-  }
-}
-
-// 设置忽略文件
-function setIgnoredFiles() {
-  // 给配置文件设置文件忽略
-  const igList = [properties.ignorePluginConfig ? '.logrc' : '', ...(properties?.settings?.git || [])];
-  // 通知是否启用标记
-  EventBus.fire<{ hint: boolean }>('add-ignore', { hint: !!properties?.settings?.git?.length });
-  if (!igList.length || (igList.length === 1 && igList[0] === '')) return;
-  let result = overwriteIgnoreFilesLocally(igList, (isGitFile: string[]) => {
-    const workspaceRoot = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
-    if (isGitFile.length) {
-      if (!workspaceRoot) return false;
-      if (properties.ignoredChanges?.added.length) {
-        const skip = properties.ignoredChanges.added.filter((sk: string) => isGitFile.includes(sk));
-        execSync(`git update-index --skip-worktree ${skip.join(' ')}`, { stdio: 'ignore', cwd: workspaceRoot });
-      }
-    }
-    if (properties.ignoredChanges?.remove.length) {
-      const skip = properties.ignoredChanges.remove.filter((sk: string) => isGitTracked(sk));
-      execSync(`git update-index --no-skip-worktree ${skip.join(' ')}`, { stdio: 'ignore', cwd: workspaceRoot });
-    }
-  });
-  MergeProperties({ isGitTracked: !!result });
-  if (result) NotificationService.info(`忽略 ${igList.length > 3 ? igList.slice(0, 3).join(',') + '...' : igList.join(',')}文件跟踪！`);
-}
-
-// 设置
-function setLogrc() {
-  setIgnoredFiles();
-  resolveResult(true);
 }
